requirements.txt
boto3==1.9.211
botocore==1.12.211
Django==2.2.4

django-environ==0.4.5 
mysql-connector-python==8.0.17 
pandas==0.25.3
requests==2.22.0
kafka-python==1.4.7
scipy==1.4.1



